changed id from int to long

+++ components/core/source/com/company/itos/core/userrolelink/dao/CreateUserRoleLinkDAO.java
@@ -25,16 +25,6 @@ public class CreateUserRoleLinkDAO {
 
 			connection = DBConnection.getDBConnection();
 
-			/*
-			 * PreparedStatement preparedStatement2 =
-			 * connection.prepareStatement
-			 * ("SELECT UsersRoleLinkSEQ.nextval FROM DUAL"); ResultSet
-			 * resultSet2 = preparedStatement2.executeQuery();
-			 * 
-			 * while (resultSet2.next()) {
-			 * userRoleLinkDetail.setUserRoleLinkID(resultSet2.getInt(1)); }
-			 */
-
 			String userRoleLinkSQLStr = "INSERT INTO UserRoleLink(userRoleLinkID, username, roleID, startDate, endDate, recordStatus, versionNo)"
 					+ "VALUES(?, ?, ?, ?, ?, 'Active', 1)";
 
@@ -76,7 +66,7 @@ public class CreateUserRoleLinkDAO {
 			ResultSet resultSet2 = preparedStatement2.executeQuery();
 
 			while (resultSet2.next()) {
-				userRoleLinkDetail.setUserRoleLinkID(resultSet2.getInt(1));
+				userRoleLinkDetail.setUserRoleLinkID(resultSet2.getLong(1));
 			}
 
 			String userRoleLinkSQLStr = "INSERT INTO UserRoleLink(userRoleLinkID, username, roleID, startDate, endDate, recordStatus, versionNo)"
