+++ components/profile/source/com/company/itos/profile/personIdentity/dao/ReadPersonIdentityDAO.java
@@ -56,8 +56,8 @@ public class ReadPersonIdentityDAO {
 			ResultSet resultSet = preparedStatementpersonIdentity.executeQuery();
 
 			while (resultSet.next()) {
-				personIdentityDetail.setPersonIdentityID(resultSet.getInt("personIdentityID"));
-				personIdentityDetail.setPersonID(resultSet.getInt("personID"));
+				personIdentityDetail.setPersonIdentityID(resultSet.getLong("personIdentityID"));
+				personIdentityDetail.setPersonID(resultSet.getLong("personID"));
 				personIdentityDetail.setAlternateID(resultSet.getString("alternateID"));
 				personIdentityDetail.setPrimaryInd(resultSet.getString("primaryInd"));
 				personIdentityDetail.setTypeCode(resultSet.getString("typeCode"));
